Fix type hint

+++ libs/plot.py
@@ -63,7 +63,7 @@ def plot_losses(data: list[float], path: pathlib.Path) -> None:
     _plot(data, path=path, plot_config=LOSS_PLOT_CONFIG)
 
 
-def plot_rewards(data: list[int], path: pathlib.Path) -> None:
+def plot_rewards(data: list[float], path: pathlib.Path) -> None:
     _plot(data, path=path, plot_config=REWARD_PLOT_CONFIG)
